requirements.txt
requests
rich
jinja2
python-dotenv
aiohttp